change: Split UserCommand and UsersCommand into two modules
