+++ src/gmp/commands/dashboards.ts
@@ -5,7 +5,7 @@
 
 import {v4 as uuid} from 'uuid';
 import HttpCommand from 'gmp/commands/http';
-import {GetSettingsResponse} from 'gmp/commands/users';
+import {GetSettingsResponse} from 'gmp/commands/user';
 import Response from 'gmp/http/response';
 import {XmlMeta} from 'gmp/http/transform/fastxml';
 import logger from 'gmp/log';
