+++ src/gmp/gmp.ts
@@ -47,7 +47,8 @@ import {ReportCommand, ReportsCommand} from 'gmp/commands/reports';
 import {RoleCommand, RolesCommand} from 'gmp/commands/roles';
 import {TaskCommand, TasksCommand} from 'gmp/commands/tasks';
 import TrashCanCommand from 'gmp/commands/trashcan';
-import {UserCommand, UsersCommand} from 'gmp/commands/users';
+import UserCommand from 'gmp/commands/user';
+import UsersCommand from 'gmp/commands/users';
 import WizardCommand from 'gmp/commands/wizard';
 import GmpSettings from 'gmp/gmpsettings';
 import GmpHttp from 'gmp/http/gmp';
