+++ src/web/pages/user-settings/FilterSettings.tsx
@@ -6,7 +6,7 @@
 import {useEffect, useMemo, useState} from 'react';
 import {useDispatch} from 'react-redux';
 
-import {DEFAULT_FILTER_SETTINGS} from 'gmp/commands/users';
+import {DEFAULT_FILTER_SETTINGS} from 'gmp/commands/user';
 import Filter, {ALL_FILTER} from 'gmp/models/filter';
 import {isDefined} from 'gmp/utils/identity';
 import Select from 'web/components/form/Select';
