+++ src/web/pages/user-settings/useSettingSave.ts
@@ -6,7 +6,7 @@
 import {useCallback, useState} from 'react';
 import {showSuccessNotification} from '@greenbone/opensight-ui-components-mantinev7';
 import {useDispatch} from 'react-redux';
-import {transformSettingName} from 'gmp/commands/users';
+import {transformSettingName} from 'gmp/commands/user';
 import useGmp from 'web/hooks/useGmp';
 import useTranslation from 'web/hooks/useTranslation';
